refactor: 폴더 구조 변경 및 불필요 파일 제거
