+++ src/components/cloud/CloudBadge.tsx
