--- src/components/domain/cloud/CloudListPage.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import CloudList from './CloudList';
-
-function CloudListPage() {
-  return <CloudList />;
-}
-
-export default CloudListPage;
